Integrate LSS slave into CANopenNode more directly. - CO_LSSslave: move expensive code from CAN receive (interrupt) to mainline   CO_LSSslave_process() function. - LSS slave now runs in parallel to other CANopen objects. - LSS slave and master can run both on same device. - LSS slave, LSS master and gateway-ascii(CiA309) LSS functions tested. - LSSusage.md updated.

+++ 305/CO_LSSmaster.h
@@ -105,7 +105,7 @@ typedef enum {
  * LSS master object.
  */
 typedef struct{
-    uint16_t         timeout_us;       /**< LSS response timeout in us */
+    uint32_t         timeout_us;       /**< LSS response timeout in us */
 
     uint8_t          state;            /**< Node is currently selected */
     uint8_t          command;          /**< Active command */
